requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
dnspython==2.6.1
idna==3.7
requests==2.32.3
soupsieve==2.5
sqlparse==0.5.1
urllib3==2.2.2
